removed the need for initResources()

+++ trunk/Visualization_Library_SDK/src/vl/Time.cpp
@@ -125,7 +125,7 @@ Real Time::currentTime()
     if (has_timer)
     {
       QueryPerformanceCounter( &PerformanceCount );
-      return (Real)((double)PerformanceCount.QuadPart/Frequency.QuadPart  - mStartTime);
+      return (Real)PerformanceCount.QuadPart/Frequency.QuadPart  - mStartTime;
     }
     else
     {
